[FilterJIN-6] 이미지뷰어 관리 ImageViewManager에서 관리
